refactor(Turn): Grid를 Turn과 TileGrid로 분리
Turn은 사용자가 갖는 게임 기회 1회, 기회를 소모한 후의 게임 상태
TileGrid는 현재 진행중인 게임의 결과 상태
