+++ src/main/java/wordle/model/GameEndTurnAction.java
@@ -2,17 +2,17 @@ package wordle.model;
 
 public class GameEndTurnAction extends GameAction {
 
-	private Grid grid;
+	private Turn turn;
 	private Word answer;
 
-	public GameEndTurnAction(Grid grid, Word answer) {
+	public GameEndTurnAction(Turn turn, Word answer) {
 		super(GameStatus.END_TURN);
-		this.grid = grid;
+		this.turn = turn;
 		this.answer = answer;
 	}
 
-	public Grid getGrid() {
-		return grid;
+	public Turn getTurn() {
+		return turn;
 	}
 
 	public Word getAnswer() {
